requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
sentence-transformers>=2.2.2
accelerate>=0.25.0
bitsandbytes>=0.41.0

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.2
openai>=1.6.0

chromadb>=0.4.18
faiss-cpu>=1.7.4
qdrant-client>=1.7.0

PyPDF2>=3.0.1
pdfplumber>=0.10.3
pytesseract>=0.3.10
pdf2image>=1.16.3

moviepy>=1.0.3
opencv-python>=4.8.1
opencv-contrib-python>=4.8.1
ffmpeg-python>=0.2.0

whisper>=1.1.10
openai-whisper>=20231117
librosa>=0.10.1
soundfile>=0.12.1

python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.2
beautifulsoup4>=4.12.2
lxml>=4.9.3

spacy>=3.7.2
nltk>=3.8.1
gensim>=4.3.2
networkx>=3.2.1
python-louvain>=0.16

scikit-learn>=1.3.2
numpy>=1.26.2
pandas>=2.1.4
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.18.0

neo4j>=5.15.0
redis>=5.0.1
pymongo>=4.6.1

celery>=5.3.4
flower>=2.0.1

python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

pillow>=10.1.0
requests>=2.31.0
aiohttp>=3.9.1
httpx>=0.25.2

pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.1

mlflow>=2.9.2
wandb>=0.16.1
tensorboard>=2.15.1

python-magic>=0.4.27
filetype>=1.2.0
tqdm>=4.66.1
rich>=13.7.0
loguru>=0.7.2

gradio>=4.8.0
streamlit>=1.29.0

redis-om>=0.2.1
sqlalchemy>=2.0.23
alembic>=1.13.0
asyncpg>=0.29.0
